chatbot/urls.py
"""
URL configuration for chatbot API endpoints.
"""

from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import DocumentViewSet, JSONDatabaseViewSet, ChatViewSet

# Create router and register viewsets
router = DefaultRouter()
router.register(r'documents', DocumentViewSet, basename='document')
router.register(r'json-database', JSONDatabaseViewSet, basename='json-database')
router.register(r'chat', ChatViewSet, basename='chat')

urlpatterns = [
    path('', include(router.urls)),
]
