requirements.txt
# Django and REST Framework
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL support (optional)

# AI/ML Libraries
transformers==4.36.2
sentence-transformers==3.0.0
huggingface-hub==0.20.3
torch==2.1.2
faiss-cpu==1.7.4  # Use faiss-gpu if GPU available

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
python-magic==0.4.27
Pillow==10.2.0

# Data Handling
numpy==1.26.3
pandas==2.1.4

# API and Async
requests==2.31.0
aiohttp==3.9.1

# Development
ipython==8.19.0
pytest==7.4.4
pytest-django==4.7.0
